requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyJWT==2.9.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
